+++ sdk/storage/azure-storage-files-datalake/src/datalake_path_client.cpp
@@ -11,6 +11,7 @@
 #include <azure/storage/common/storage_per_retry_policy.hpp>
 #include <azure/storage/common/storage_retry_policy.hpp>
 
+#include "azure/storage/files/datalake/datalake_constants.hpp"
 #include "azure/storage/files/datalake/datalake_utilities.hpp"
 #include "azure/storage/files/datalake/version.hpp"
 
@@ -248,12 +249,35 @@ namespace Azure { namespace Storage { namespace Files { namespace DataLake {
     protocolLayerOptions.Permissions = options.Permissions;
     auto result = Details::DataLakeRestClient::Path::Create(
         m_dfsUri, *m_pipeline, options.Context, protocolLayerOptions);
-    auto ret = Models::CreatePathResult();
+    Models::CreatePathResult ret;
     ret.ETag = std::move(result->ETag.GetValue());
     ret.LastModified = std::move(result->LastModified.GetValue());
     ret.ContentLength = std::move(result->ContentLength);
     return Azure::Core::Response<Models::CreatePathResult>(
         std::move(ret), result.ExtractRawResponse());
+  }
+
+  Azure::Core::Response<Models::CreatePathResult> PathClient::CreateIfNotExists(
+      Models::PathResourceType type,
+      const CreatePathOptions& options) const
+  {
+    try
+    {
+      auto createOptions = options;
+      createOptions.AccessConditions.IfNoneMatch = ETagWildcard;
+      return Create(type, createOptions);
+    }
+    catch (StorageException& e)
+    {
+      if (e.ErrorCode == Details::DataLakePathAlreadyExists)
+      {
+        Models::CreatePathResult ret;
+        ret.Created = false;
+        return Azure::Core::Response<Models::CreatePathResult>(
+            std::move(ret), std::move(e.RawResponse));
+      }
+      throw;
+    }
   }
 
   Azure::Core::Response<Models::DeletePathResult> PathClient::Delete(
@@ -267,8 +291,34 @@ namespace Azure { namespace Storage { namespace Files { namespace DataLake {
     protocolLayerOptions.IfModifiedSince = options.AccessConditions.IfModifiedSince;
     protocolLayerOptions.IfUnmodifiedSince = options.AccessConditions.IfUnmodifiedSince;
     protocolLayerOptions.RecursiveOptional = options.Recursive;
-    return Details::DataLakeRestClient::Path::Delete(
+    auto result = Details::DataLakeRestClient::Path::Delete(
         m_dfsUri, *m_pipeline, options.Context, protocolLayerOptions);
+    Models::DeletePathResult ret;
+    ret.ContinuationToken = std::move(result->ContinuationToken);
+    ret.Deleted = true;
+    return Azure::Core::Response<Models::DeletePathResult>(
+        std::move(ret), result.ExtractRawResponse());
+  }
+
+  Azure::Core::Response<Models::DeletePathResult> PathClient::DeleteIfExists(
+      const DeletePathOptions& options) const
+  {
+    try
+    {
+      return Delete(options);
+    }
+    catch (StorageException& e)
+    {
+      if (e.ErrorCode == Details::DataLakeFilesystemNotFound
+          || e.ErrorCode == Details::DataLakePathNotFound)
+      {
+        Models::DeletePathResult ret;
+        ret.Deleted = false;
+        return Azure::Core::Response<Models::DeletePathResult>(
+            std::move(ret), std::move(e.RawResponse));
+      }
+      throw;
+    }
   }
 
   Azure::Core::Response<Models::GetPathPropertiesResult> PathClient::GetProperties(
@@ -332,7 +382,7 @@ namespace Azure { namespace Storage { namespace Files { namespace DataLake {
     {
       acl = Models::Acl::DeserializeAcls(result->Acl.GetValue());
     }
-    auto ret = Models::GetPathAccessControlResult{};
+    Models::GetPathAccessControlResult ret;
     ret.ETag = std::move(result->ETag);
     ret.LastModified = std::move(result->LastModified);
     if (!acl.HasValue())